Finished the Bet Option Modal, sitll unsure about the ref
